requirements.txt
amqp==2.5.2
asgiref==3.2.5
billiard==3.6.3.0
celery==4.4.2
Django==3.0.4
djangorestframework==3.11.0
importlib-metadata==1.5.0
kombu==4.6.8
pytz==2019.3
redis==3.4.1
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
